requirements.txt
streamlit==1.38.0
streamlit-authenticator==0.3.1
sqlalchemy==2.0.34
pyodbc==5.1.0
pandas==2.2.2
numpy==1.26.4
plotly==5.23.0
PyYAML==6.0.2
python-dotenv==1.0.1   
cryptography==43.0.1 